# Skip useless handler mappings calls while connecting