--- a/src/Controls/src/Core/Label/Label.Mapper.cs
+++ b/src/Controls/src/Core/Label/Label.Mapper.cs
@@ -21,7 +21,7 @@ namespace Microsoft.Maui.Controls
 
 			// these are really a single property
 			LabelHandler.Mapper.ReplaceMapping<Label, ILabelHandler>(nameof(Text), MapText);
-			LabelHandler.Mapper.ReplaceMapping<Label, ILabelHandler>(nameof(FormattedText), MapText);
+			LabelHandler.Mapper.ReplaceMapping<Label, ILabelHandler>(nameof(FormattedText), MapFormattedText);
 
 			LabelHandler.Mapper.ReplaceMapping<Label, ILabelHandler>(nameof(LineBreakMode), MapLineBreakMode);
 			LabelHandler.Mapper.ReplaceMapping<Label, ILabelHandler>(nameof(MaxLines), MapMaxLines);
@@ -54,8 +54,17 @@ namespace Microsoft.Maui.Controls
 			MapTextOrFormattedText(handler, label);
 		static void MapTextTransform(ILabelHandler handler, Label label) =>
 			MapTextOrFormattedText(handler, label);
+		static void MapFormattedText(ILabelHandler handler, Label label)
+		{
+			if (label.IsConnectingHandler()) return;
+
+			MapText(handler, label);
+		}
+
 		static void MapTextOrFormattedText(ILabelHandler handler, Label label)
 		{
+			if (label.IsConnectingHandler()) return;
+
 			if (label.HasFormattedTextSpans)
 				handler.UpdateValue(nameof(FormattedText));
 			else