--- a/src/Core/src/Handlers/Element/ElementHandler.cs
+++ b/src/Core/src/Handlers/Element/ElementHandler.cs
@@ -2,7 +2,7 @@
 
 namespace Microsoft.Maui.Handlers
 {
-	public abstract partial class ElementHandler : IElementHandler
+	public abstract partial class ElementHandler : IElementHandler, IElementHandlerStateExhibitor
 	{
 		public static IPropertyMapper<IElement, IElementHandler> ElementMapper = new PropertyMapper<IElement, IElementHandler>()
 		{
@@ -15,6 +15,9 @@ namespace Microsoft.Maui.Handlers
 		internal readonly IPropertyMapper _defaultMapper;
 		internal readonly CommandMapper? _commandMapper;
 		internal IPropertyMapper _mapper;
+		ElementHandlerState _handlerState;
+
+		ElementHandlerState IElementHandlerStateExhibitor.State => _handlerState;
 
 		protected ElementHandler(IPropertyMapper mapper, CommandMapper? commandMapper = null)
 		{
@@ -40,24 +43,38 @@ namespace Microsoft.Maui.Handlers
 			_ = view ?? throw new ArgumentNullException(nameof(view));
 
 			if (VirtualView == view)
+			{
 				return;
+			}
 
 			var oldVirtualView = VirtualView;
 
 			bool setupPlatformView = oldVirtualView == null;
 
 			VirtualView = view;
-			PlatformView ??= CreatePlatformElement();
+			if (PlatformView is null)
+			{
+				_handlerState = ElementHandlerState.Connecting;
+				PlatformView = CreatePlatformElement();
+			}
+			else
+			{
+				_handlerState = ElementHandlerState.Reconnecting;
+			}
 
 			if (VirtualView.Handler != this)
+			{
 				VirtualView.Handler = this;
+			}
 
 			// We set the previous virtual view to null after setting it on the incoming virtual view.
 			// This makes it easier for the incoming virtual view to have influence
 			// on how the exchange of handlers happens.
 			// We will just set the handler to null ourselves as a last resort cleanup
 			if (oldVirtualView?.Handler != null)
+			{
 				oldVirtualView.Handler = null;
+			}
 
 			if (setupPlatformView)
 			{
@@ -77,6 +94,8 @@ namespace Microsoft.Maui.Handlers
 			}
 
 			_mapper.UpdateProperties(this, VirtualView);
+
+			_handlerState = ElementHandlerState.Connected;
 		}
 
 		public virtual void UpdateValue(string property)
@@ -129,6 +148,8 @@ namespace Microsoft.Maui.Handlers
 				PlatformView = null;
 				DisconnectHandler(oldPlatformView);
 			}
+
+			_handlerState = ElementHandlerState.Disconnected;
 		}
 	}
 }