--- a/src/Core/src/Handlers/View/ViewHandler.Windows.cs
+++ b/src/Core/src/Handlers/View/ViewHandler.Windows.cs
@@ -112,6 +112,8 @@ namespace Microsoft.Maui.Handlers
 		{
 			if (view is IContextFlyoutElement contextFlyoutContainer)
 			{
+				if (handler.IsConnectingHandler() && contextFlyoutContainer.ContextFlyout is null) return;
+
 				MapContextFlyout(handler, contextFlyoutContainer);
 			}
 		}