--- a/src/Core/src/Handlers/View/ViewHandler.cs
+++ b/src/Core/src/Handlers/View/ViewHandler.cs
@@ -26,17 +26,19 @@ namespace Microsoft.Maui.Handlers
 		/// A dictionary that maps the virtual view properties to their platform view counterparts.
 		/// </summary>
 		public static IPropertyMapper<IView, IViewHandler> ViewMapper =
-#if ANDROID
-			// Use a custom mapper for Android which knows how to batch the initial property sets
-			new AndroidBatchPropertyMapper<IView, IViewHandler>(ElementHandler.ElementMapper)
-#else
 			new PropertyMapper<IView, IViewHandler>(ElementHandler.ElementMapper)
-#endif
 			{
 				// This property is a special one and needs to be set before other properties.
 				[nameof(IViewHandler.ContainerView)] = MapContainerView,
 #if ANDROID
-				[AndroidBatchPropertyMapper.InitializeBatchedPropertiesKey] = MapInitializeBatchedProperties,
+				// On Android we try to reduce JNI calls by batch-initializing many properties in one call.
+				// Properties involved are:
+				//     Visibility, MinimumHeight, MinimumWidth, IsEnabled, Opacity, TranslationX, TranslationY,
+				//     Scale, ScaleX, ScaleY, Rotation, RotationX, RotationY, AnchorX, AnchorY
+				//
+				// The single mappers will behave as noop thanks to the `handler.IsConnectingHandler()` check.
+				// The end user can still replace the mappers or append code to them, and it will be executed.
+				["_InitializeBatchedProperties"] = MapInitializeBatchedProperties,
 #endif
 
 				[nameof(IView.AutomationId)] = MapAutomationId,
@@ -247,6 +249,16 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapMinimumHeight(IViewHandler handler, IView view)
 		{
+#if ANDROID
+			if (handler.IsConnectingHandler())
+			{
+				// Mapped through _InitializeBatchedProperties
+				return;
+			}
+#else
+			if (handler.IsConnectingHandler() && double.IsNaN(view.MinimumHeight)) return;
+#endif
+
 			((PlatformView?)handler.PlatformView)?.UpdateMinimumHeight(view);
 		}
 
@@ -257,6 +269,8 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapMaximumHeight(IViewHandler handler, IView view)
 		{
+			if (handler.IsConnectingHandler() && double.IsNaN(view.MaximumHeight)) return;
+
 			((PlatformView?)handler.PlatformView)?.UpdateMaximumHeight(view);
 		}
 
@@ -267,6 +281,16 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapMinimumWidth(IViewHandler handler, IView view)
 		{
+#if ANDROID
+			if (handler.IsConnectingHandler())
+			{
+				// Mapped through _InitializeBatchedProperties
+				return;
+			}
+#else
+			if (handler.IsConnectingHandler() && double.IsNaN(view.MinimumWidth)) return;
+#endif
+
 			((PlatformView?)handler.PlatformView)?.UpdateMinimumWidth(view);
 		}
 
@@ -277,6 +301,8 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapMaximumWidth(IViewHandler handler, IView view)
 		{
+			if (handler.IsConnectingHandler() && double.IsNaN(view.MaximumWidth)) return;
+
 			((PlatformView?)handler.PlatformView)?.UpdateMaximumWidth(view);
 		}
 
@@ -287,6 +313,14 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapIsEnabled(IViewHandler handler, IView view)
 		{
+#if ANDROID
+			if (handler.IsConnectingHandler())
+			{
+				// Mapped through _InitializeBatchedProperties
+				return;
+			}
+#endif
+
 			((PlatformView?)handler.PlatformView)?.UpdateIsEnabled(view);
 		}
 
@@ -297,6 +331,16 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapVisibility(IViewHandler handler, IView view)
 		{
+#if ANDROID
+			if (handler.IsConnectingHandler())
+			{
+				// Mapped through _InitializeBatchedProperties
+				return;
+			}
+#else
+			if (handler.IsConnectingHandler() && view.Visibility == Visibility.Visible) return;
+#endif
+
 			if (handler.HasContainer)
 				((PlatformView?)handler.ContainerView)?.UpdateVisibility(view);
 
@@ -333,6 +377,8 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapFlowDirection(IViewHandler handler, IView view)
 		{
+			if (handler.IsConnectingHandler() && view.FlowDirection == FlowDirection.MatchParent) return;
+
 			((PlatformView?)handler.PlatformView)?.UpdateFlowDirection(view);
 		}
 
@@ -343,6 +389,16 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapOpacity(IViewHandler handler, IView view)
 		{
+#if ANDROID
+			if (handler.IsConnectingHandler())
+			{
+				// Mapped through _InitializeBatchedProperties
+				return;
+			}
+#else
+			if (handler.IsConnectingHandler() && view.Opacity == 1) return;
+#endif
+
 			if (handler.HasContainer)
 			{
 				((PlatformView?)handler.ContainerView)?.UpdateOpacity(view);
@@ -360,6 +416,8 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapAutomationId(IViewHandler handler, IView view)
 		{
+			if (handler.IsConnectingHandler() && view.AutomationId is null) return;
+
 			((PlatformView?)handler.PlatformView)?.UpdateAutomationId(view);
 		}
 
@@ -370,7 +428,13 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapClip(IViewHandler handler, IView view)
 		{
-			handler.UpdateValue(nameof(IViewHandler.ContainerView));
+			if (handler.IsConnectingHandler() && view.Clip is null) return;
+
+			if (!handler.IsMappingProperties())
+			{
+				// ContainerView is already being mapped
+				handler.UpdateValue(nameof(IViewHandler.ContainerView));
+			}
 
 			((PlatformView?)handler.ContainerView)?.UpdateClip(view);
 		}
@@ -382,7 +446,13 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapShadow(IViewHandler handler, IView view)
 		{
-			handler.UpdateValue(nameof(IViewHandler.ContainerView));
+			if (handler.IsConnectingHandler() && view.Shadow is null) return;
+
+			if (!handler.IsMappingProperties())
+			{
+				// ContainerView is already being mapped
+				handler.UpdateValue(nameof(IViewHandler.ContainerView));
+			}
 
 			((PlatformView?)handler.ContainerView)?.UpdateShadow(view);
 		}