--- a/src/Core/src/Handlers/View/ViewHandler.cs
+++ b/src/Core/src/Handlers/View/ViewHandler.cs
@@ -269,7 +269,7 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapMaximumHeight(IViewHandler handler, IView view)
 		{
-			if (handler.IsConnectingHandler() && double.IsNaN(view.MaximumHeight)) return;
+			if (handler.IsConnectingHandler() && double.IsPositiveInfinity(view.MaximumHeight)) return;
 
 			((PlatformView?)handler.PlatformView)?.UpdateMaximumHeight(view);
 		}
@@ -301,7 +301,7 @@ namespace Microsoft.Maui.Handlers
 		/// <param name="view">The associated <see cref="IView"/> instance.</param>
 		public static void MapMaximumWidth(IViewHandler handler, IView view)
 		{
-			if (handler.IsConnectingHandler() && double.IsNaN(view.MaximumWidth)) return;
+			if (handler.IsConnectingHandler() && double.IsPositiveInfinity(view.MaximumWidth)) return;
 
 			((PlatformView?)handler.PlatformView)?.UpdateMaximumWidth(view);
 		}