… src/Core/src/Handlers/View/AndroidBatchPropertyMapper.cs
… src/Core/src/Handlers/View/ViewHandler.iOS.cs