### Description of Change

This is an alternative initial version of #27042.
Everything has already been explained there.

**Based on/includes #28077 PR**

**Credits to @MartyIX to have explored the effectiveness of skipping specific mappers**.

Key differences from the other PR:
- Avoid pre-mapping defaults: if the developer replaced the mapper and we skip execution, we're also skipping their custom code which would bring them wonder why their code it's not working while it should be
- Adds easy-to-use extension methods to `IElement` and `IElementHandler` to get the handler state
- We can selectively skip native calls close to them and by platform (for example on iOS `MappingFrame` takes care of all the coordinates, even translations, while on other platforms the situation is different)

### Benchmark using Nalu Weather app ###

https://github.com/nalu-development/nalu/tree/main/Samples/Nalu.Maui.Weather

Click on the links in the table to download the speedscope.json file.

| Platform | `SetVirtualView` (SR4) | `SetVirtualView` (PR) | Improvement |
|-------------------------- |----------:|---------:|--------------:|
| iOS | [411.99ms](https://github.com/user-attachments/files/19012310/ios-maui-9-sr4.speedscope.json) | [233.62ms](https://github.com/user-attachments/files/19012307/ios-pr.speedscope.json) | **43%** |
| Android | [340.62ms](https://github.com/user-attachments/files/19012497/android-maui-9-sr4.speedscope.json) | [299.34ms](https://github.com/user-attachments/files/19012496/android-pr.speedscope.json) | **12%** |
| Windows | [1.94s](https://github.com/user-attachments/files/19024175/windows-maui-9-sr4.speedscope.json) |[1.50s](https://github.com/user-attachments/files/19024194/windows-pr.speedscope.json) | **22%** |

Thanks to @MartyIX for collecting speedscopes on Windows!

### Benchmark (from CI - on iOS UI tests)

Note: there is some variability in the benchmark caused by the runner itself, but looking at many recent PRs, `main` branch never goes under 2600ms.

**`main` branch**

![image](https://github.com/user-attachments/assets/81a3397a-efb3-4d39-ac38-016421c10a5c)

**#27042 original PR**

![image](https://github.com/user-attachments/assets/19518b94-34d9-4361-bd17-dab201cab674)

**this `PR`**
![image](https://github.com/user-attachments/assets/42d9ef0e-17f8-4eeb-8c2b-14f9082c4e2c)


### Issues Fixed

Fixes #17303