@marcojak 2025-01-30
Definitely it would be good to try it.

Great to see attention to the performances as MAUI struggles a lot there.