@bcaceiro 2025-01-30
@albyrock87 just tried the latest nightly, everything seems smooth! ( only tested so far on Android)