@bcaceiro 2025-01-30
@PureWeen not really, since my app is quite complex (ble, views, etc) I haven't still recorded a speedscope, so it is really  just more a "feeling" than any measurable data. In debug mode, and release mode, it seems the loads are a bit faster in pages Appearing, not dramatically faster, but seems faster.

Once I have opportunity I will try to record something though