@albyrock87 2025-01-30
@bcaceiro if you record the speedscopes for comparison, you should check timings on `SetVirtualView` method.