@bcaceiro 2025-02-04
Update: this is not available on nightlies. So I really didn' test this PR in my previous "feeling".
 Nevertheless, now I have sucessfully configured manually the downloaded Nuget, and yes, now it is noticeable in a page with a collectionview, labels and 2 buttons, the difference in the render ( in debug mode). I haven't recorded a speedscope since I'm having some issues with VSCode.